Fix for 'bug' that was showing private records to users that should not have view access.
All authenticated user are added to the Contributor group at the
data_record_repository level so that they can add to dataset_records to
the repository. This is then inherited in the dataset_records meaning
that all authenticated users could see private records. This commit
removes view access from the Contributor groups for private and pending
review states.