RIFCS modifications: 1 bug fix and 1 feature add.
Extra newlines in the data location text box were resulting in empty
<addressPart></addressPart> elements being generated in the RIFCS - this
is invalid RIFCS and was resulting in records being marked as invalid at
the RDA end after harvest. Have added check for empty lines at the point
of RIFCS generation.

Have added code to check for a cc licence in the licence field and if
found a related link to the cc licence document is added to the RIFCS.