Updates to RIF-CS Activity record generation
Project start and end dates have been included in the description note
and the temporal coverage element has been removed from the record.

The temporal coverage was incorrectly using the project start year as a
coverage start date - this is not what the temporal coverage dates were
intended for. In RIF-CS v1.3 existance date elements have been added -
when we move to this, the start and end dates will need to be used for
this.